server/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autopep8==1.6.0
bandit==1.7.4
bidict==0.22.0
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
multidict==6.0.2
pbr==5.9.0
pycodestyle==2.9.1
pynput==1.7.6
pyobjc-core==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Quartz==8.5
python-engineio==4.3.4
python-socketio==5.7.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
smmap==5.0.0
stevedore==3.5.0
toml==0.10.2
urllib3==1.26.11
yarl==1.8.1
